+agrego managers de seguridad, temperatura y simulacion +agrego query para devices x tipo -agrego funcionalidad al actuador de temperatura (!falta obtener el parametro extra y validar si puede o no calentar y enfriar)

+++ trunk/server/src/server/model/devices/Sensor.java
@@ -17,7 +17,7 @@ public abstract class Sensor extends Device {
 		return mState;
 	}
 	
-	public int getValue() {
+	public float getValue() {
 		ArduinoHandler.getInstance().addOperation(this, false);
 		//TODO make async
 		return 0;
